Some changes with Data and Functionality of the Filing Cabinet
Made it so folders did not get activated whenever clicking in specific spots on closed cabinets.
Made it so cabinets could not be interacted with while a file was open.
Encapsulated the function that handles the functionality of the filing cabinet, so that it did not get called on accident.